Big Gameplay Changes
What's new?
GameState - Simple enum class that record the state of the game depending on the ability in use
GameController - Anything that controlls game flow will go in here. Currently it has a static reference to a gamestate variable that will be shared by all scripts
Object - New implementations for orbiting a player
PlayerController - pretty much stayed the same